Improved support.
At least I hope I did.
Added support for ES6 Class prototype functions.
Fixed bug in some cases where a module would be loaded from cache, but the cache was for the filename entry that we use to store which files require which files. I think it might be important info to store, we could possibly do away with it though if we never have to update those modules that call another when that module changes.
Yeah im not making sense